adding default case for LanguageIndexingConfiguration class
