+++ wikitongues/wikitongues/config/indexing.cfg
@@ -51,4 +51,4 @@ base_id : {your base id}
 api_key : {your api key}
 # DO NOT MODIFY
 fake : false
-max_records : 10
+max_records
